+++ akshare/__init__.py
@@ -2081,9 +2081,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.60 add: add stock_telegraph_cls interface
 1.6.61 fix: fix futures_shfe_warehouse_receipt interface
 1.6.62 fix: fix futures_zh_spot interface
+1.6.63 fix: fix fund_portfolio_em and futures_roll_yield interface
 """
 
-__version__ = "1.6.62"
+__version__ = "1.6.63"
 __author__ = "AKFamily"
 
 import sys