+++ akshare/stock_feature/stock_hist_em.py
@@ -1692,7 +1692,7 @@ if __name__ == "__main__":
         symbol="301183",
         period="weekly",
         start_date="20220516",
-        end_date="20220522",
+        end_date="20220722",
         adjust="hfq",
     )
     print(stock_zh_a_hist_df)